requirements.txt
python-telegram-bot==20.3
selenium==4.10.0
